# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Films API",
    "version": "1.0.0",
    "description": "API para la gestión de películas"
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Servidor de desarrollo"
    }
  ],
  "paths": {
    "/api/films": {
      "get": {
        "summary": "Lista todas las películas",
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Film"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/film/:id": {
      "get": {
        "summary": "Obtiene detalles de una película específica",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Film"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/film": {
      "post": {
        "summary": "Inserta una nueva película",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewFilm"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Película insertada exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/film/update/:id": {
      "put": {
        "summary": "Modifica una película",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewFilm"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Película modificada exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/film/delete/:id": {
      "delete": {
        "summary": "Elimina una película",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Película eliminada exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Film": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "year": {
            "type": "integer"
          },
          "rating": {
            "type": "number"
          },
          "director": {
            "type": "string"
          }
        }
      },
      "NewFilm": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "year": {
            "type": "integer"
          },
          "rating": {
            "type": "number"
          },
          "director": {
            "type": "integer"
          }
        }
      }
    }
  }
}